e-learning/backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-browser-reload==1.18.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
